==> notebooks/01_convert_txt_to_csv.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import random

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

random.seed(42)

PATH_TO_DATA = '../data/'


### Slice verbatim.txt and convert to CSV

original_verbatim_file = PATH_TO_DATA + 'raw/verbatim.txt'

original_verbatim_df = pd.read_csv(original_verbatim_file, delimiter='\t', low_memory=False)
original_verbatim_df.head(5)

relevant_columns = ['gbifID', 'eventDate', 'continent', 'country', 'stateProvince', 'decimalLatitude', 'decimalLongitude', 'scientificName', 'vernacularName', 'higherClassification', 'family']
original_verbatim_df_slice = original_verbatim_df[relevant_columns]
# original_verbatim_df_slice.to_csv(PATH_TO_DATA + 'processed/single_csv/verbatim.csv', index=False)
original_verbatim_df_slice.head(5)

##### Display moths in original dataset

moths = original_verbatim_df[original_verbatim_df['family'] == 'Tineidae']
print(moths.shape)
moths.head(5)

### Slice multimedia.txt and convert to CSV

original_multimedia_file = PATH_TO_DATA + 'raw/multimedia.txt'
original_multimedia_df =  pd.read_csv(original_multimedia_file, delimiter='\t', low_memory=False)

relevant_columns = ['gbifID', 'format', 'identifier']
original_multimedia_df_slice = original_multimedia_df[relevant_columns]
# original_multimedia_df_slice.to_csv(PATH_TO_DATA + 'processed/single_csv/multimedia.csv', index=False)
original_multimedia_df_slice.head(5)
