==> notebooks/03_create_datasets.ipynb <==
import pandas as pd
import random

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

random.seed(42)

PATH_TO_DATA = '../data/'

verbatim_file = PATH_TO_DATA + 'processed/single_csv/verbatim.csv'
verbatim_df = pd.read_csv(verbatim_file)
verbatim_df.tail(3)

multimedia_file = PATH_TO_DATA + 'processed/single_csv/multimedia.csv'
multimedia_df = pd.read_csv(multimedia_file)
multimedia_df.tail(3)

species_counts = verbatim_df['scientificName'].value_counts()

### Creating testing dataset (top20max50)
- 20 species with most occurrences
- 50 images per species
- Total images: 1000

species_for_testing_ds = species_counts.head(20).keys()
species_for_testing_ds

gbifID_blacklist = [1966132886, 4459520332] # entering gbifIDs which match images with poor quality
selected_species = species_counts.head(20).keys()
selected_data = {'gbifID': [], 'scientificName': []}

for species in selected_species:
    species_ids = verbatim_df[verbatim_df['scientificName'] == species]['gbifID'].tolist()

    valid_species_ids = [id for id in species_ids if id not in gbifID_blacklist]
    
    if len(valid_species_ids) >= 50:
        selected = random.sample(species_ids, 50)
    
    # Füge die ausgewählten IDs zur Liste hinzu
    for gbifID in selected:
        selected_data['gbifID'].append(gbifID)
        selected_data['scientificName'].append(species)


selected_df = pd.DataFrame(selected_data)

output_file = PATH_TO_DATA + 'processed/testing_dataset_top20max50.csv'
# selected_df.to_csv(output_file, index=False)

### Creating testing dataset (top589max3000)
- 589 species with most occurrences (all above 1k)
- maximum of 3000 images per species (randomly selected)
- Total images: ~1.3 million

full_df = pd.merge(verbatim_df, multimedia_df, on='gbifID')
print(full_df.shape)
full_df.tail(3)

full_df = full_df.drop_duplicates(subset=['gbifID']) # reduce to single occ per gbifID
print(full_df.shape)

full_df_grouped = full_df.groupby('scientificName') # Group by 'scientificName'
filtered_dfs = []

MAX_AMOUNT = 3000
MIN_AMOUNT = 1000

for name, group in full_df_grouped:
    sample_size = len(group)
    if sample_size >= MAX_AMOUNT:
        sampled = group.sample(n=MAX_AMOUNT, random_state=42) # classes with more than 5000 samples get reduced to 5000
        filtered_dfs.append(sampled)
    elif MIN_AMOUNT <= sample_size < MAX_AMOUNT: # classes between 1000 and 5000 stay unchanged
        filtered_dfs.append(group)
    # Classes with less than 1000 samples are skipped

# Concatenate the filtered DataFrames
final_df = pd.concat(filtered_dfs)
print(final_df.shape)
final_df['scientificName'].value_counts()

# save dataset to file
# full_df.to_csv(PATH_TO_DATA + 'processed/dataset_full.csv', index=False)
# final_df.to_csv(PATH_TO_DATA + 'processed/dataset_top589_max3000.csv', index=False)

### Create Datasets for CV and Baseline

top589_max3000_file = PATH_TO_DATA + 'processed/resized_datasets/dataset_top589_max3000.csv'
top589_max3000_df = pd.read_csv(top589_max3000_file)

STATES_TO_IGNORE = ['ERROR', 'MISSING', 'BLACK']
top589_max3000_df = top589_max3000_df[~top589_max3000_df['status'].isin(STATES_TO_IGNORE)]
print(top589_max3000_df.shape)
top589_max3000_df.head(3)


top589_max3000_df['scientificName'].value_counts().values



dataset_configs = {
    'top277': (277, [3000, 2000, 1000, 500]), # (amount of classes, [amount of samples per species])
    'top387': (387, [2000, 1000, 500]),
    'top589': (589, [1000, 500])
}

for class_amount, sample_amounts in dataset_configs.values():
    selected_species = top589_max3000_df['scientificName'].value_counts().head(class_amount).keys() # first 277 classes have 3000 samples


    df = top589_max3000_df[top589_max3000_df['scientificName'].isin(selected_species)]
    df_grouped = df.groupby('scientificName') # Group by 'scientificName'

    for amount in sample_amounts:
        filtered_dfs = []

        for name, group in df_grouped:
            sampled = group.sample(n=min(len(group), amount), random_state=42)
            filtered_dfs.append(sampled)
        variation_df = pd.concat(filtered_dfs)
        print(f'Shape for dataset with {amount} samples for {class_amount} species: {variation_df.shape}')
        #if input(f'Enter 1 to save dataset_top{class_amount}_max{amount}.csv') == '1':
        variation_df.to_csv(PATH_TO_DATA + f'processed/resized_datasets/dataset_top{class_amount}_max{amount}.csv', index=False)
        print(f'Saved dataset_top{class_amount}_max{amount}.csv')
