==> notebooks/11_specific_accuracies.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import glob


PATH_TO_DATA = '/home/lgierz/data/'

model_names = ["Linear Classifier", "KNN"]
fm_names = ['resnet', 'dino']

dataset_configs = {
    'top277': (277, [3000, 2000, 1000, 500]),
    'top387': (387, [2000, 1000, 500]),
    'top589': (589, [1000, 500])
}
PATH_TO_CA = PATH_TO_DATA + 'run_results/confidence_tests/'

PATH_TO_IMAGES = '/mnt/data/lgierz/moth_dataset_top589_max3000/'
PATH_TO_LABELS = '/home/lgierz/BA_MothClassification/data_big_files/dataset_full_top589_max3000.csv'
PATH_TO_CORRECTED_CONFS = PATH_TO_CA + 'all_confidences_results.csv'

labels_csv = pd.read_csv(PATH_TO_LABELS)
confidences_csv = pd.read_csv(PATH_TO_CORRECTED_CONFS)

# create real mapping for labels_csv
species = labels_csv['scientificName']
unique_species = sorted(species.unique())
species_mapping = {species: idx for idx, species in enumerate(unique_species)}

print(unique_species)
print(species_mapping)
print(len(unique_species))

# Create mappings from species name to family and from species name to genus
species2family = labels_csv[['scientificName','family']].drop_duplicates().set_index('scientificName')['family'].to_dict()

label_family_mapping = {}
label_genus_mapping = {}
for sp, lbl in species_mapping.items():
    label_family_mapping[lbl] = species2family.get(sp, 'Unknown')
    label_genus_mapping[lbl] = sp.split(' ')[0]

print(len(set(label_family_mapping.values())), ' ', end='')
print(label_family_mapping)

print(len(set(label_genus_mapping.values())), ' ', end='')
print(label_genus_mapping)

confidences_csv['FamilyCorrect'] = confidences_csv.apply(
    lambda row: label_family_mapping[row['Label']] == label_family_mapping[row['PredictedLabel']],
    axis=1
)

confidences_csv['GenusCorrect'] = confidences_csv.apply(
    lambda row: label_genus_mapping[row['Label']] == label_genus_mapping[row['PredictedLabel']],
    axis=1
)



confidences_csv['Top5Correct'] = confidences_csv.apply(
    lambda row: row['Label'] in [lab for lab, _ in eval(row['Confidence'])[:5]],
    axis=1
)

### Calculate Top1 Accuracy


acc = confidences_csv.groupby(['FoundationalModel', 'Model', 'Species', 'Samples'])['Correct'].mean().apply(lambda x: round(x*100, 2)).reset_index()
acc[((acc['Species'] == 277) & ((acc['Samples'] == 1000) | (acc['Samples'] == 3000)))  | ((acc['Species'] == 589)& (acc['Samples'] == 1000))].reset_index(drop=True)


### Calculate Top5 Accuracy


top5acc = confidences_csv.groupby(['FoundationalModel', 'Model', 'Species', 'Samples'])['Top5Correct'].mean().apply(lambda x: round(x*100, 2)).reset_index()
top5acc[((acc['Species'] == 277) & ((acc['Samples'] == 1000) | (acc['Samples'] == 3000)))  | ((acc['Species'] == 589)& (acc['Samples'] == 1000))].reset_index(drop=True)

### Genus Accuracy


genusacc = confidences_csv.groupby(['FoundationalModel', 'Model', 'Species', 'Samples'])['GenusCorrect'].mean().apply(lambda x: round(x*100, 2)).reset_index()
genusacc[((acc['Species'] == 277) & ((acc['Samples'] == 1000) | (acc['Samples'] == 3000)))  | ((acc['Species'] == 589)& (acc['Samples'] == 1000))].reset_index(drop=True)

### Family Accuracies


familyacc = confidences_csv.groupby(['FoundationalModel', 'Model', 'Species', 'Samples'])['FamilyCorrect'].mean().apply(lambda x: round(x*100, 2)).reset_index()
familyacc[((acc['Species'] == 277) & ((acc['Samples'] == 1000) | (acc['Samples'] == 3000)))  | ((acc['Species'] == 589)& (acc['Samples'] == 1000))].reset_index(drop=True)
